test1/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-core==1.4.1
django-cors-headers==3.10.0
django-mysql==4.2.0
djangorestframework==3.12.4
mysql-client==0.0.1
mysqlclient==2.1.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
